+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * nostr: expose NIP49 `log_n` ([DanConwayDev])
 * nostr: add tags indexes to `Event` ([Yuki Kishimoto])
 * nostr: add `hex::decode_to_slice` ([Yuki Kishimoto])
+* nostr: add `SecretKey::generate` ([Yuki Kishimoto])
 * pool: add `RelayPool::start` ([Yuki Kishimoto])
 * pool: add `NegentropyDirection` default ([Yuki Kishimoto])
 * sdk: add `Client::builder()` ([Yuki Kishimoto])